+++ backend/model.py
@@ -3,8 +3,8 @@ import math
 
 class Model:
     def __init__(self, ):
-        self.df = {}
-        self.tdfi = {}
+        self.df = {}  # Key:(Term) Value: (# of docs in which Term appears)
+        self.tdfi = {} # Main Key:(FilePath), Value: Has Dict with Keys: 'index'(term_freq_index for that file). 'total_terms_count', 'last_modified' of file
 
     def to_dict(self):
         return {
@@ -18,6 +18,19 @@ class Model:
         model.df = data.get('df', {})
         model.tdfi = data.get('tdfi', {})
         return model
+
+    def remove_document(self, file_path):
+        if file_path in self.tdfi:
+            for key in self.tdfi[file_path]['index'].keys():
+                if key in self.df:
+                    count = self.df[key]
+                    self.df[key] = (count - 1)
+
+
+    def model_check(self):
+        total_documents = len(self.tdfi)
+        for value in self.df.values():
+            assert value <= total_documents, "Term doc counts should be <= Total Docs"
 
 
 def compute_tf(term, num_of_terms_in_doc, term_freq_dict):
